+++ src/views/text-match.usage.js
@@ -3,7 +3,27 @@ export default {
         view: 'text-match',
         data: {
             text: 'I am matched text or a text with matches!',
-            match: /(match)/
+            match: /match/
         }
-    }
+    },
+    examples: [
+        {
+            title: 'Raw string as match',
+            demo: {
+                view: 'text-match',
+                data: {
+                    text: 'I am matched text or a text with matches!',
+                    match: 'match'
+                }
+            }
+        },
+        {
+            title: 'Using with a link',
+            demo: {
+                view: 'link',
+                data: { href: '#' },
+                content: 'text-match:{ text: "Test link testtest", match: /test/i }'
+            }
+        }
+    ]
 };