+++ src/views/tree.js
@@ -7,6 +7,7 @@ export default function(discovery) {
             limit = data.length;
         }
 
+        container.classList.add('incomplete');
         data
             .slice(offset, offset + limit)
             .reduce(
@@ -44,22 +45,20 @@ export default function(discovery) {
                 }),
                 Promise.resolve()
             )
-            .then(() =>
-                discovery.view.maybeMoreButtons(
-                    container,
-                    null,
-                    data.length,
-                    offset + limit,
-                    limit,
-                    (offset, limit) => renderTreeLines(container, renderStack, data, context, offset, limit)
-                )
-            );
+            .then(() => discovery.view.maybeMoreButtons(
+                container,
+                null,
+                data.length,
+                offset + limit,
+                limit,
+                (offset, limit) => renderTreeLines(container, renderStack, data, context, offset, limit)
+            ) || container.classList.remove('incomplete'));
     }
 
-    function buildTreeLines(data, context, childrenGetter, expanded) {
-        function processChildren(array, expanded, popCount = 0) {
+    function buildTreeLines(data, context, itemConfig, expanded) {
+        function processChildren(array, expanded, itemConfig, popCount = 0) {
             array.forEach((data, index, array) => {
-                const children = discovery.query(childrenGetter, data, context);
+                const children = discovery.query(itemConfig.children, data, context);
                 const hasChildren = Array.isArray(children) && children.length > 0;
                 const last = index === array.length - 1;
                 const leafExpanded =
@@ -75,7 +74,7 @@ export default function(discovery) {
                     expanded: leafExpanded,
                     last,
                     hasChildren,
-                    children: leafExpanded ? null : childrenGetter,
+                    children: leafExpanded ? null : itemConfig.children,
                     shift: last && (!leafExpanded || !hasChildren) ? popCount + 1 : 0
                 });
 
@@ -83,6 +82,7 @@ export default function(discovery) {
                     processChildren(
                         children,
                         typeof expanded === 'number' ? expanded - 1 : expanded,
+                        discovery.view.composeConfig(itemConfig, itemConfig.itemConfig),
                         last ? popCount + 1 : 0
                     );
                 }
@@ -92,7 +92,7 @@ export default function(discovery) {
         const leafs = [];
         const visited = new Set();
 
-        processChildren(data, expanded);
+        processChildren(data, expanded, itemConfig);
 
         return leafs;
     }
@@ -115,7 +115,7 @@ export default function(discovery) {
             expanded = typeof expanded === 'function' ? expanded : discovery.view.listLimit(expanded, 1);
 
             if (limitLines) {
-                const lines = buildTreeLines(data, context, children, expanded);
+                const lines = buildTreeLines(data, context, this.composeConfig({ children }, itemConfig), expanded);
                 const renderStack = {
                     container: el,
                     itemConfig: this.composeConfig({